+++ conda_smithy/ci_register.py
@@ -488,6 +488,28 @@ def travis_configure(user, project):
             response.raise_for_status()
 
 
+def add_token_to_travis(user, project):
+    """Add the BINSTAR_TOKEN to travis."""
+
+    headers = travis_headers()
+
+    repo_info = travis_get_repo_info(user, project)
+    repo_id = repo_info["id"]
+
+    data = {
+        "env_var.name": "BINSTAR_TOKEN",
+        "env_var.value": anaconda_token,
+        "env_var.public": "false",
+    }
+    r = requests.post(
+        "{}/repo/{repo_id}/env_vars".format(travis_endpoint, repo_id=repo_id),
+        headers=headers,
+        json=data,
+    )
+    if r.status_code != 201:
+        r.raise_for_status()
+
+
 def travis_cleanup(org, project):
     if os.getenv("GH_TRAVIS_TOKEN"):
         gh = github.Github(os.getenv("GH_TRAVIS_TOKEN"))