+++ conda_smithy/lint_recipe.py
@@ -453,33 +453,41 @@ def lintify(meta, recipe_dir=None, conda_forge=False):
                 )
                 continue
 
-    # 23: non noarch builds shouldn't use version constraints on python
+    # 23: non noarch builds shouldn't use version constraints on python and r-base
+    check_languages = ["python", "r-base"]
     host_reqs = requirements_section.get("host") or []
     run_reqs = requirements_section.get("run") or []
-    if build_section.get("noarch") is None and not outputs_section:
-        filtered_host_reqs = [
-            req
-            for req in host_reqs
-            if req == "python" or req.startswith("python ")
-        ]
-        filtered_run_reqs = [
-            req
-            for req in run_reqs
-            if req == "python" or req.startswith("python ")
-        ]
-        if filtered_host_reqs and not filtered_run_reqs:
-            lints.append(
-                "If python is a host requirement, it should be a run requirement."
-            )
-        for reqs in [filtered_host_reqs, filtered_run_reqs]:
-            if "python" in reqs:
-                continue
-            for req in reqs:
-                constraint = req.split(" ", 1)[1]
-                if constraint.startswith(">") or constraint.startswith("<"):
-                    lints.append(
-                        "Non noarch: python packages should have a python requirement without any version constraints."
+    for language in check_languages:
+        if build_section.get("noarch") is None and not outputs_section:
+            filtered_host_reqs = [
+                req
+                for req in host_reqs
+                if req.partition(" ")[0] == str(language)
+            ]
+            filtered_run_reqs = [
+                req
+                for req in run_reqs
+                if req.partition(" ")[0] == str(language)
+            ]
+            if filtered_host_reqs and not filtered_run_reqs:
+                lints.append(
+                    "If {0} is a host requirement, it should be a run requirement.".format(
+                        str(language)
                     )
+                )
+            for reqs in [filtered_host_reqs, filtered_run_reqs]:
+                if str(language) in reqs:
+                    continue
+                for req in reqs:
+                    constraint = req.split(" ", 1)[1]
+                    if constraint.startswith(">") or constraint.startswith(
+                        "<"
+                    ):
+                        lints.append(
+                            "Non noarch packages should have {0} requirement without any version constraints.".format(
+                                str(language)
+                            )
+                        )
 
     # hints
     # 1: suggest pip